requirements.txt
Flask
Flask-SQLAlchemy
thefuzz
python-Levenshtein
requests
beautifulsoup4
